pyproject.toml
[project]
name = "time-serie-explo"
version = "0.1.0"
description = "Exploration of time series forecasting with CPU/CUDA/XPU support"
readme = "README.md"
authors = [
    { name = "Your Name", email = "you@example.com" }
]
requires-python = ">=3.10,<3.13"

# Base dependencies (without torch - managed separately per backend)
dependencies = [
    "numpy>=1.24.0,<2.0.0",
    "pandas>=2.0.0",
    "scipy>=1.10.0",
    "scikit-learn>=1.3.0",
    "pyyaml>=6.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "plotly>=5.17.0",
    "streamlit>=1.28.0",
    "pytorch-lightning>=2.0.0",
    "einops>=0.7.0",
    "darts>=0.32.0",
    "statsmodels>=0.14.0",
    "shap>=0.42.0",
    "timeshap>=1.0.0",
    "captum>=0.7.0",
    "optuna>=3.0.0",
    "tqdm>=4.65.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "codecarbon>=2.3.0",
    "tensorboard>=2.13.0",
    "optuna-dashboard>=0.13.0",
    "mlflow>=3.8.0",
    "pastas>=1.0.0",
    "sqlparse>=0.4.0",
    "numba>=0.57.0",
]

[project.optional-dependencies]
cpu = [
    "torch>=2.5.0",
    "torchaudio>=2.5.0",
    "torchvision>=0.20.0",
]
cuda = [
    "torch>=2.5.0",
    "torchaudio>=2.5.0",
    "torchvision>=0.20.0",
]
xpu = [
    "torch>=2.5.0",
    "torchaudio>=2.5.0",
    "torchvision>=0.20.0",
    "intel-extension-for-pytorch>=2.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["dashboard"]
